Finish fixing various errors. [Assets-v0.4]
Now is moderately functional. It appears to work.

+++ src/World.cpp
@@ -7,7 +7,7 @@
 using namespace Nimbus;
 
 World::World(Ogre::SceneManager* sceneManager):
-	worldBounds(Ogre::Real(250), Ogre::Real(250))
+	worldBounds(Ogre::Real(1000), Ogre::Real(1000))
 {
 	idCounter = 0;
 	this->mSceneManager = sceneManager;
